Likes bug fixed

+++ app/api/post-card/route.ts
@@ -2,6 +2,23 @@ import { db } from "@/db";
 import { and, eq, sql } from "drizzle-orm";
 import { NextResponse } from "next/server";
 import { posts, users, likes } from "@/db/schema";
+import { Linkedin } from "lucide-react";
+export async function GET(req: Request) {
+    const { searchParams } = new URL(req.url);
+    const userId = searchParams.get("userId");
+    const postId = searchParams.get("postId");
+
+    if (!userId || !postId) return NextResponse.json({liked: false});
+
+    const row = await db.query.likes.findFirst({
+        where: (likes, { eq, and }) => and(
+            eq(likes.postID, Number(postId)),
+            eq(likes.userID, Number(userId))
+        )
+    })
+
+    return NextResponse.json({ liked: !!row })
+}
 
 export async function POST(req: Request) {
     const body = await req.json();
@@ -31,21 +48,37 @@ export async function POST(req: Request) {
                 await db.update(posts)
                     .set({ likes: sql`${posts.likes} + 1` })
                     .where(eq(posts.id, numPostId));
+            } else {
+                // return NextResponse.json({message: "Post is already liked.", liked : true, updated: false}, {status: 200});
+                return NextResponse.json({liked: true, updated: false})
             }
         } else if (action === "unlike") {
             const numUserId = Number(userId);
             const numPostId = Number(postId);
 
-            await db.delete(likes).where(
-                and(
-                    eq(likes.userID, numUserId),
-                    eq(likes.postID, numPostId)
+            const row = await db.query.likes.findFirst({
+                where: (likes, { eq, and }) => (
+                    and(
+                        eq(likes.postID, numPostId),
+                        eq(likes.userID, numUserId)
+                    )
                 )
-            );
+            })
 
-            await db.update(posts)
-                .set({ likes: sql`${posts.likes} - 1` })
-                .where(eq(posts.id, numPostId));
+            if (row) {
+                await db.delete(likes).where(
+                    and(
+                        eq(likes.userID, numUserId),
+                        eq(likes.postID, numPostId)
+                    )
+                );
+
+                await db.update(posts)
+                    .set({ likes: sql`GREATEST(0, ${posts.likes} - 1)` })
+                    .where(eq(posts.id, numPostId));
+            } else {
+                return NextResponse.json({liked: false, updated: false})
+            }
         } else if (repost !== undefined) {
             const numPostId = Number(id);
             await db.update(posts)
